backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
requests==2.31.0
django-cors-headers==4.3.1
gunicorn>=21.2.0
whitenoise>=6.6.0
